refactor: apply layering to member domain
