+++ src/main/java/backend/knowhow/domain/auth/dto/request/RoleRequest.java
@@ -1,6 +1,6 @@
 package backend.knowhow.domain.auth.dto.request;
 
-import backend.knowhow.domain.auth.domain.Role;
+import backend.knowhow.domain.member.domain.Role;
 
 public record RoleRequest(
         Role role
