+++ src/main/java/backend/knowhow/domain/auth/service/AuthService.java
@@ -1,10 +1,10 @@
 package backend.knowhow.domain.auth.service;
 
-import backend.knowhow.domain.auth.domain.Member;
-import backend.knowhow.domain.auth.domain.Role;
+import backend.knowhow.domain.member.domain.Member;
+import backend.knowhow.domain.member.domain.Role;
 import backend.knowhow.domain.auth.dto.response.AuthResponse;
 import backend.knowhow.domain.auth.dto.response.KakaoUserInfo;
-import backend.knowhow.domain.auth.repository.MemberRepository;
+import backend.knowhow.domain.member.repository.MemberRepository;
 import backend.knowhow.domain.auth.repository.RefreshTokenRepository;
 import backend.knowhow.global.common.exception.BaseException;
 import backend.knowhow.global.common.response.ErrorType;
