+++ src/main/java/backend/knowhow/domain/member/domain/GuardianLink.java
@@ -1,4 +1,4 @@
-package backend.knowhow.domain.auth.domain;
+package backend.knowhow.domain.member.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
