+++ src/main/java/backend/knowhow/domain/member/domain/Role.java
@@ -1,4 +1,4 @@
-package backend.knowhow.domain.auth.domain;
+package backend.knowhow.domain.member.domain;
 
 public enum Role {
     DRIVER,
