+++ src/main/java/backend/knowhow/domain/member/dto/response/MemberInfoResponse.java
@@ -1,4 +1,4 @@
-package backend.knowhow.domain.auth.dto.response;
+package backend.knowhow.domain.member.dto.response;
 
 public record MemberInfoResponse(
         Long id,
